build: implementing reset-pyenv script for Linux. Works fine for python 3.9
