+++ build.py
@@ -1,141 +1,29 @@
-#!/usr/bin/env python3
-
-# https://stackoverflow.com/questions/60073711/how-to-build-c-extensions-via-poetry
-from distutils.sysconfig import  get_python_inc, get_python_lib
-from typing import List, Union
-
-import os
-import os.path as path
-import subprocess
-import sys
-import platform
 from pathlib import Path
 
-from distutils.errors import CCompilerError, DistutilsExecError, DistutilsPlatformError
-
-# TODO dvp: check if the issue with build requirements is resolved and poetry updated
-# see https://github.com/python-poetry/poetry/pull/2794
-# import numpy as np  # numpy is in build requirements, so, it should be available
-# workaround
-# try:
-#     import numpy as np
-# except ImportError:
-#     subprocess.check_call("poetry run python -m pip install mkl-devel numpy".split())
-#     import numpy as np
-import numpy as np
-
-from build_nlopt import NLOptBuildExtension, NLOptBuild
-from setuptools import Extension
-
-# see https://habr.com/ru/post/210450/
+from extension_builder import MCKitBuilder
+from extension_geometry import geometry_extension
+from extension_nlopt import NLOptBuildExtension
 from setuptools.dist import Distribution
-
-PLATFORM_SYSTEM = platform.system()
 
 
 class BinaryDistribution(Distribution):
-    def is_pure(self):
+    def is_pure(self):  # noqa
         return False
 
 
-def get_dirs(environment_variable):
-    dirs = os.environ.get(environment_variable, "")
-
-    if dirs:
-        dirs = dirs.split(os.pathsep)
-    else:
-        dirs = []
-
-    return dirs
-
-
-def insert_directories(
-    destination: List[str], value: Union[str, List[str]]
-) -> List[str]:
-    dirs = []
-    if isinstance(value, list):
-        dirs.extend(value)
-    elif value not in destination:
-        dirs.append(value)
-    for old in destination:
-        if old not in dirs:
-            dirs.append(old)
-    return dirs
-
-
-include_dirs = get_dirs("INCLUDE_PATH")
-
-include_dirs = insert_directories(include_dirs, np.get_include())
-include_dirs = insert_directories(include_dirs, get_python_inc())
-
-library_dirs = get_dirs("LIBRARY_PATH")
-library_dirs = insert_directories(library_dirs, os.path.join(sys.prefix, "libs"))  # for PythonXX.dll
-
-geometry_dependencies = [
-    "mkl_rt",
-    "nlopt",
-]
-
-if PLATFORM_SYSTEM == "Linux":
-    include_dirs = insert_directories(include_dirs, path.join(sys.prefix, "include"))
-    library_dirs = insert_directories(library_dirs, path.join(sys.prefix, "lib"))
-    extra_compile_args = ["-O3", "-w"]
-elif PLATFORM_SYSTEM == "Windows":
-    include_dirs = insert_directories(include_dirs, os.path.join(sys.prefix, "Library/include"))
-    library_dirs = insert_directories(library_dirs, os.path.join(sys.prefix, "Library/lib"))
-    extra_compile_args = ["/O2"]
-else:
-    raise EnvironmentError(f"Not implemented for platform {PLATFORM_SYSTEM}")
-
-
-geometry_sources = [
-    path.join("mckit", "src", src)
-    for src in ["geometrymodule.c", "box.c", "surface.c", "shape.c", "rbtree.c"]
-]
-
-
-ext_modules = [
-    NLOptBuildExtension(
-        "nlopt"
-    ),
-    Extension(
-        "mckit.geometry",
-        sources=geometry_sources,
-        include_dirs=include_dirs,
-        libraries=geometry_dependencies,
-        library_dirs=library_dirs,
-        extra_compile_args=extra_compile_args,
-        language="c",
-    ),
-]
-
-
-# class BuildFailed(Exception):
-#     pass
-#
-#
-# class ExtBuilder(build_ext):
-#     def run(self):
-#         try:
-#             build_ext.run(self)
-#         except (DistutilsPlatformError, FileNotFoundError):
-#             raise BuildFailed("File not found. Could not compile C extension.")
-#
-#     def build_extension(self, ext):
-#         try:
-#             build_ext.build_extension(self, ext)
-#         except (CCompilerError, DistutilsExecError, DistutilsPlatformError, ValueError):
-#             raise BuildFailed("Could not compile C extension.")
+ext_modules = [NLOptBuildExtension(), geometry_extension]
 
 
 def build(setup_kwargs):
     """
-    This function is mandatory in order to build the extensions.
+    Set specific distribution options.
+
+    This function is called with setup.py generated by pip from pyproject.toml.
     """
     setup_kwargs.update(
         {
             "ext_modules": ext_modules,
-            "cmdclass": {"build_ext": NLOptBuild},
+            "cmdclass": {"build_ext": MCKitBuilder},
             "package_data": {"mckit": ["data/isotopes.dat", "nlopt.dll"]},
             "distclass": BinaryDistribution,
             "setup_requires": [
@@ -146,5 +34,3 @@ def build(setup_kwargs):
             "long_description": Path("README.rst").read_text(encoding="utf8"),
         }
     )
-
-
